:white_check_mark: Move fixture definition to conftest.py
